refactor: move Navigator to under Visualizer/Widget/Navigator
