+++ src/components/molecules/Visualizer/Widget/Navigator/NavigatorPresenter/index.stories.tsx
