+++ src/components/molecules/Visualizer/Widget/Navigator/NavigatorPresenter/index.test.tsx
