+++ src/components/molecules/Visualizer/Widget/Navigator/NavigatorPresenter/utils.ts
