+++ src/components/molecules/Visualizer/Widget/Navigator/index.tsx
@@ -1,16 +1,16 @@
-import NavigatorUI from "@reearth/components/atoms/Navigator";
 import { ComponentProps as WidgetProps } from "@reearth/components/molecules/Visualizer/Widget";
 
 import { SceneProperty } from "../../Engine";
 
 import useHooks from "./hooks";
+import NavigatorPresenter from "./NavigatorPresenter";
 
 export type Props = WidgetProps<SceneProperty>;
 
 const Navigator = ({ sceneProperty }: Props): JSX.Element | null => {
   const { degree, events } = useHooks({ sceneProperty });
 
-  return <NavigatorUI degree={degree} {...events} />;
+  return <NavigatorPresenter degree={degree} {...events} />;
 };
 
 export default Navigator;
